@@ -10,7 +10,7 @@
 	"exports": {
 		".": {
 			"import": "./built/esm/index.js",
-			"types": "./built/dts/index.ts"
+			"types": "./built/dts/index.d.ts"
 		},
 		"./*": {
 			"import": "./built/esm/*",
@@ -36,16 +36,16 @@
 		"test": "npm run jest"
 	},
 	"devDependencies": {
-		"@microsoft/api-extractor": "7.37.0",
+		"@microsoft/api-extractor": "7.38.0",
 		"@types/jest": "29.5.5",
-		"@types/node": "20.6.2",
-		"@types/seedrandom": "3.0.5",
-		"@types/uuid": "9.0.4",
-		"@typescript-eslint/eslint-plugin": "6.7.2",
-		"@typescript-eslint/parser": "6.7.2",
+		"@types/node": "20.8.4",
+		"@types/seedrandom": "3.0.6",
+		"@types/uuid": "9.0.5",
+		"@typescript-eslint/eslint-plugin": "6.7.5",
+		"@typescript-eslint/parser": "6.7.5",
 		"chalk": "5.3.0",
 		"copyfiles": "2.4.1",
-		"eslint": "8.49.0",
+		"eslint": "8.51.0",
 		"eslint-plugin-import": "2.28.1",
 		"jest": "29.7.0",
 		"peggy": "3.0.2",
