@@ -3,6 +3,11 @@ import vue from '@vitejs/plugin-vue'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-	base: '/aiscript/',
-  plugins: [vue()]
+	base: './',
+  plugins: [vue()],
+	server: {
+		fs: {
+			allow: [ '..' ]
+		}
+	},
 })
