@@ -299,8 +299,8 @@ describe('Infix expression', () => {
 	test.concurrent('syntax symbols vs infix operators', async () => {
 		const res = await exe(`
 		<: match true {
-			1 == 1 => "true"
-			1 < 1 => "false"
+			case 1 == 1 => "true"
+			case 1 < 1 => "false"
 		}
 		`);
 		eq(res, STR('true'));
@@ -313,8 +313,8 @@ describe('Infix expression', () => {
 	test.concurrent('number + match expression', async () => {
 		const res = await exe(`
 			<: 1 + match 2 == 2 {
-				true => 3
-				false  => 4
+				case true => 3
+				case false => 4
 			}
 		`);
 		eq(res, NUM(4));
@@ -324,17 +324,11 @@ describe('Infix expression', () => {
 		eq(await exe('<: eval { 1 } + eval { 1 }'), NUM(2));
 	});
 
-	test.concurrent('disallow line break', async () => {
-		try {
-			await exe(`
+	test.concurrent('allow line break', async () => {
+		eq(await exe(`
 			<: 1 +
 			1 + 1
-			`);
-		} catch (e) {
-			assert.ok(true);
-			return;
-		}
-		assert.fail();
+		`), NUM(3));
 	});
 
 	test.concurrent('escaped line break', async () => {
@@ -1494,9 +1488,9 @@ describe('match', () => {
 	test.concurrent('Basic', async () => {
 		const res = await exe(`
 		<: match 2 {
-			1 => "a"
-			2 => "b"
-			3 => "c"
+			case 1 => "a"
+			case 2 => "b"
+			case 3 => "c"
 		}
 		`);
 		eq(res, STR('b'));
@@ -1505,9 +1499,9 @@ describe('match', () => {
 	test.concurrent('When default not provided, returns null', async () => {
 		const res = await exe(`
 		<: match 42 {
-			1 => "a"
-			2 => "b"
-			3 => "c"
+			case 1 => "a"
+			case 2 => "b"
+			case 3 => "c"
 		}
 		`);
 		eq(res, NULL);
@@ -1516,10 +1510,10 @@ describe('match', () => {
 	test.concurrent('With default', async () => {
 		const res = await exe(`
 		<: match 42 {
-			1 => "a"
-			2 => "b"
-			3 => "c"
-			* => "d"
+			case 1 => "a"
+			case 2 => "b"
+			case 3 => "c"
+			default => "d"
 		}
 		`);
 		eq(res, STR('d'));
@@ -1528,13 +1522,13 @@ describe('match', () => {
 	test.concurrent('With block', async () => {
 		const res = await exe(`
 		<: match 2 {
-			1 => 1
-			2 => {
+			case 1 => 1
+			case 2 => {
 				let a = 1
 				let b = 2
 				(a + b)
 			}
-			3 => 3
+			case 3 => 3
 		}
 		`);
 		eq(res, NUM(3));
@@ -1544,7 +1538,7 @@ describe('match', () => {
 		const res = await exe(`
 		@f(x) {
 			match x {
-				1 => {
+				case 1 => {
 					return "ai"
 				}
 			}
